Rename home App component and fix import
